refactor: extract styles from screen
