+++ src/components/atoms/Loading/styles.tsx
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+	box: {
+		justifyContent: 'center',
+		backgroundColor: 'black',
+	},
+	animation: {
+		width: '70%',
+		display: 'flex',
+		alignSelf: 'center',
+		justifyContent: 'center',
+	},
+});
