+++ src/components/atoms/Splash/styles.tsx
@@ -1,0 +1,16 @@
+import {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+	box: {
+		flex: 1,
+		width: '100%',
+		justifyContent: 'center',
+		backgroundColor: '#000',
+	},
+	animation: {
+		width: '60%',
+		display: 'flex',
+		flexDirection: 'column',
+		alignSelf: 'center',
+	},
+});
