+++ src/components/screens/Home/styles.tsx
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		flex: 1,
+		backgroundColor: '#000',
+	},
+	headerRight: {flexDirection: 'row'},
+});
