+++ src/components/screens/Search/index.tsx
@@ -1,9 +1,9 @@
 import React, {useEffect, useState} from 'react';
-import {FlatList, SafeAreaView, StyleSheet, View} from 'react-native';
+import {FlatList, SafeAreaView, View} from 'react-native';
 import {SearchBar, Text} from 'react-native-elements';
-import {getSearchingMovies, getUpcomingMovies} from '../../utils';
-import List from '../atoms/List';
-
+import {getSearchingMovies, getUpcomingMovies} from '../../../utils';
+import List from '../../atoms/List';
+import styles from './styles';
 interface Props {
 	navigation: any;
 	route: any;
@@ -83,32 +83,3 @@ const Search: React.FC<Props> = ({navigation}) => {
 };
 
 export default Search;
-
-const styles = StyleSheet.create({
-	searchBarContainerStyle: {height: 55, backgroundColor: '#000'},
-	searchBarInputContainerStyle: {backgroundColor: '#393e42', height: 35},
-	MainContainer: {
-		flex: 1,
-		backgroundColor: 'white',
-	},
-
-	item: {
-		paddingLeft: 15,
-		paddingTop: 8,
-		paddingBottom: 8,
-	},
-
-	ListHeaderComponentStyle: {
-		borderBottomColor: '#000',
-		borderBottomWidth: 0,
-	},
-	topSearch: {
-		color: '#fff',
-		marginHorizontal: 10,
-		marginBottom: 10,
-		fontWeight: '900',
-		fontSize: 20,
-	},
-
-	headerRight: {flexDirection: 'row'},
-});
